fixed: the gif layout had broken the binary slide layout, in a way i didn't think possible...
i defined what i thought was a locally-scoped width variable specific
to the gif case of changeSlide(), but i guess it was scoped to the
function, not the case. maybe you can't scope to case; maybe i'm a
fool! is this related to that "declaration hoisting" thing i've heard
of?